+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All breaking changes are marked with [BC] and potentially require API consumer c
 - Replace `{Horizontal,Vertical}Layout` with `{H,V}Layout` which are short-lived and implement Widget. [BC]
 - Make `Table` scroll properly when window is too small to show whole table.
 - Add `BehaviorContext` to `TableRow` trait. This allows passing parameters to the `behavior` of a column. [BC]
+- Make `layout_linearly` take `weights` parameter. [BC]
+- Require weights for nodes in `HSplit` and `VSplit`. [BC]
 
 ## [0.2.5] - 2020-11-14
 ### Fixed